[optimisation]: optimised weather functionality
